periutza cu săpun!!!

+++ src/main/java/com/parking/parkinglot/common/CarDto.java
@@ -9,19 +9,19 @@ import java.util.Objects;
  * DTO for {@link com.parking.parkinglot.entities.Car}
  */
 public class CarDto {
-    Long id;
+    Integer id;
     String licensePlate;
     String parkingSpot;
     String ownerName;
 
-    public CarDto(String ownerName, String licensePlate, Long id, String parkingSpot) {
+    public CarDto(String ownerName, String licensePlate, Integer id, String parkingSpot) {
         this.ownerName = ownerName;
         this.licensePlate = licensePlate;
         this.id = id;
         this.parkingSpot = parkingSpot;
   }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
